feat: 기물 움직임에 대한 검증 책임을 board에서 game이 갖도록 별도의 클래스로 분리
- PieceMoveVerifier가 전적으로 움직임에 대한 책임을 검증함
- Board에서 findPiece조회를 해야하는 부분을 FunctionalInterface를 적용
